covid21 module works whrough update_everything()
