+++ python/covid21/Fit.py
@@ -9,6 +9,10 @@ import matplotlib.dates as mdates
 import matplotlib.pyplot as plt
 import os
 import pyreadr
+import glob
+#import re
+import statistics as stats
+from tabulate import tabulate
 
 
 
@@ -180,7 +184,7 @@ class Fit(GG.Geography):
             ax[2].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
 
             if (show_median):
-                med = median(np.exp(log_beta))
+                med = stats.median(np.exp(log_beta))
                 logmed = np.log(med)
                 medstr = '%s = %.3g'%('$\\tilde{\\beta}$',med)
                 ax[2].text(ax[2].get_xlim()[0],logmed,medstr,ha='left',va='bottom',fontsize=10)
@@ -231,7 +235,7 @@ class Fit(GG.Geography):
             ax[3].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
 
             if (show_median):
-                med = median(np.exp(self.diag['logmu']))
+                med = stats.median(np.exp(self.diag['logmu']))
                 logmed = np.log(med)
                 medstr = '%s = %.3g'%('$\\tilde{\\mu}$',med)
                 ax[3].text(ax[3].get_xlim()[0],logmed,medstr,ha='left',va='bottom',fontsize=10)
@@ -291,4 +295,250 @@ class Fit(GG.Geography):
             plt.pause(3)
         else:
             plt.show()
+
+def get_mu_atD1(ext='.RData',fit_files = []):
+    if (len(fit_files) < 1):
+        fit_files = glob.glob(cv.fit_path+'*'+ext)
+    else:
+        for i,ff in enumerate(fit_files):
+            fit_files[i] = cv.fit_path+fit_files[i]+ext
+
+    mud_cols = ['moniker','lag','logmu','mu']
+    tmp = np.empty(shape=(len(fit_files),4),dtype=object)
+    for i,ff in enumerate(fit_files):
+        moniker = os.path.splitext(os.path.basename(ff))[0]
+        fit = Fit(ff)
+        muD1 = -1.0
+        obs_deaths = fit.diag['obs_deaths']
+        for t in range(0,len(fit.diag.index)):
+            if (muD1 < 0.0) and (obs_deaths[t] > 0.0):
+                muD1 = np.exp(fit.diag['logmu'][t])
+                lag = t-1
+                tmp[i] = (moniker,lag,fit.diag['logmu'][t],muD1)
+
+    mud = pd.DataFrame(tmp,columns=mud_cols)
+    mud = mud.sort_values(by='lag',ascending=True)
+    print(mud)
+    mud.to_csv(cv.fit_path+'mud.csv',index=False)
+    fig, ax = plt.subplots(1,figsize=(6.5,4.5))
+    ax.set_ylabel('First '+r'$\ln\ \mu\ (da^{-1})$')
+    ax.set_xlabel('Lag (days)') 
+#   ax.hist(mud['logmu'],density=True,bins=3)
+    ax.plot(mud['lag'],(mud['logmu']))
+    plt.show(True)
+
+def make_rate_plots(yvarname = 'logbeta',ext = '.RData', 
+                    fit_files = [], show_medians = False, 
+                    add_doubling_time = False, show_order_date = True, save=False):
+    print(yvarname)
+    if (yvarname == 'logbeta'):
+        ylabel =r'$\ln \beta\ (da^{-1})$'
+    elif (yvarname == 'logmu'):
+        ylabel =r'$\ln \mu\ (da^{-1})$'
+    elif (yvarname == 'gamma'):
+        ylabel =r'$\gamma\ (da^{-1})$'
+    else:
+        sys.exit('Unknown yvarname '+yvarname)
+
+    fig, ax = plt.subplots(1,figsize=(6.5,4.5))
+    if (len(fit_files) < 1):
+        suffix = '_g'
+        fit_files = glob.glob(cv.fit_path+'*'+ext)
+    #   fit_files = glob.glob(cv.fit_path+'constrainID/'+'*'+ext)
+    else:
+        suffix = '_'+str(len(fit_files))
+        for i,ff in enumerate(fit_files):
+            fit_files[i] = cv.fit_path+fit_files[i]+ext
+    for i,ff in enumerate(fit_files):
+        print(i,ff)
+        fit = Fit(ff)
+        if (i == 0):
+            GU.make_date_axis(ax)
+            ax.set_ylabel(ylabel)
+       
+        pdate = []
+        Date0 = datetime.strptime(fit.date0,'%Y-%m-%d')
+        for t in range(0,len(fit.diag.index)):
+            pdate.append(mdates.date2num(Date0 + timedelta(days=t)))
+
+        yvar = fit.diag[yvarname]
+        if (yvarname == 'gamma'):
+            print(yvarname)
+            print(min(yvar),max(yvar))
+        ax.plot(pdate,yvar)
+
+    #   sigma_logbeta is the standard deviation of the generating
+    #   random walk, NOT the standard deviation of the estimated
+    #   random effect
+        if (yvarname == 'logbeta' and len(fit_files) <=4):
+        #   sigma_logbeta = fit.get_est_or_init('logsigma_logbeta')
+        #   plot_error(ax,pdate,yvar,sigma_logbeta,logscale=True)
+            GU.plot_error(ax,pdate,yvar,fit.diag['SElogbeta'],logscale=True)
+
+        sn = GG.short_name(fit.moniker)
+        if (yvarname == 'logbeta'):
+            GU.mark_ends(ax,pdate,yvar,sn,'b')
+        else:
+            GU.mark_ends(ax,pdate,yvar,sn,'b')
+
+        if (show_medians):
+            med = stats.median(np.exp(yvar))
+            logmed = np.log(med)
+            ax.plot(ax.get_xlim(),[logmed,logmed],linewidth=1,
+                    color=ax.get_lines()[-1].get_color())
+
+    if (show_order_date):
+        GU.add_order_date(ax)
+
+#   finagle doubling time axis at same scale as beta
+    if (add_doubling_time):
+        yticks = np.arange(-7,1,1)
+        ax.set_ylim(min(yticks)-1,max(yticks)+1)
+        ax.set_yticks(yticks)
+        dtax = ax.twinx()
+        dtax.set_ylim(ax.get_ylim())
+        dtax.grid(False,axis='y') # omit grid lines
+        dtax.set_ylabel('Doubling Time (da)')
+    #   render now to get the tick positions and labels
+    #   fig.canvas.draw()
+        y2_ticks = dtax.get_yticks()
+        labels = dtax.get_yticklabels()
+        for i in range(0,len(y2_ticks)):
+            y2_ticks[i] = np.log(2)/np.exp(y2_ticks[i])
+        #   labels[i] = '%.1f'%y2_ticks[i]
+            labels[i] = round(float(y2_ticks[i]),2)
+
+        dtax.tick_params(length=0)
+        dtax.set_yticklabels(labels)
+
+    if save:
+        gfile = cv.graphics_path+yvarname+'_summary'+suffix+'.png'
+        fig.savefig(gfile)
+        print('plot saved as',gfile)
+        plt.show(block=False)
+    else:
+        plt.show(block=True)
+
+def make_fit_plots(ext = '.RData'):
+    fit_files = glob.glob(cv.fit_path+'*'+ext)
+    print('found',len(fit_files),ext,'files in',cv.fit_path)
+    plt.rc('figure', max_open_warning = 0)
+    for ff in fit_files:
+        fit = Fit(ff)
+    #   fit.print_metadata()
+        fit.plot(logscale=True)
+
+    fit = Fit(cv.fit_path+'NassauNY'+ext)
+    fit.plot(save=True,logscale=False)
+    fit = Fit(cv.fit_path+'Miami-DadeFL'+ext)
+    fit.plot(save=True,logscale=False)
+    fit = Fit(cv.fit_path+'New_York_CityNY'+ext)
+    fit.plot(save=True,logscale=False)
+    fit = Fit(cv.fit_path+'Los_AngelesCA'+ext)
+    fit.plot(save=True,logscale=False)
+
+def make_fit_table(ext = '.RData'):
+    fit_files = glob.glob(cv.fit_path+'*'+ext)
+    print('found',len(fit_files),ext,'files in',cv.fit_path)
+    # mtime = os.path.getmtime(cspath) cv.NYT_counties
+    #   dtime = datetime.fromtimestamp(mtime)
+    #   self.updated = str(dtime.date())
+    # updated from https://github.com/nytimes/covid-19-data.git
+
+#   md_cols = ['county','N0','ntime','prop_zero_deaths','fn']
+    md_cols = ['county','ntime','prop_zero_deaths','fn','C']
+    es_cols = ['logsigma_logCP','logsigma_logDP','logsigma_logbeta','logsigma_logmu',
+               'logsigma_logC','logsigma_logD','mbeta','mmu'] #,'mgamma']
+    tt_cols = md_cols + es_cols
+    header = ['County','$n$','$p_0$','$f$','$C$',
+              '$\sigma_{\eta_C}$', '$\sigma_{\eta_D}$', '$\sigma_\\beta$','$\sigma_\\mu$',
+              '$\sigma_{\ln I}$','$\sigma_{\ln D}$','$\\tilde{\\beta}$','$\\tilde{\\mu}$']
+            #,'$\\tilde\\gamma$']
+
+    tt = pd.DataFrame(columns=tt_cols,dtype=None)
+
+    func = pd.DataFrame(columns=['fn'])
+#   mgamma = pd.DataFrame(columns=['mgamma'])
+    mbeta = pd.DataFrame(columns=['mbeta'])
+    mmu = pd.DataFrame(columns=['mmu'])
+    sigfigs = 3
+#   for ff in fit_files:
+    for k in range(0,len(fit_files)):
+        ff = fit_files[k]
+    #   fn = ff.replace(' ','_',5) 
+    #   pn = fit_path+fn+'.RData'
+    #   fit=pyreadr.read_r(pn)
+        print('adding fit',k,ff)
+        fit = Fit(ff)
+        ests  = fit.ests #['ests']
+        meta = fit.md #['meta']
+        diag = fit.diag #['diag']
+        row = pd.Series(index=tt_cols)
+        county = fit.get_metadata_item('county')  
+        row['county'] = GG.pretty_county(county)
+        for k in range(1,len(tt_cols)):
+            n = tt_cols[k]
+            v = fit.get_est_or_init(n)
+            if ("logsigma" in n):
+                if (v != None):
+                    v = float(np.exp(v))
+            row.iloc[k] = v
+
+    #   row['N0'] = int(get_metadata('N0',meta))
+        row['C'] = fit.get_metadata_item('convergence')
+        row['ntime'] = int(fit.get_metadata_item('ntime'))
+    #   row['prop_zero_deaths'] = round(float(fit.get_metadata_item('prop_zero_deaths')),sigfigs)
+        row['prop_zero_deaths'] = float(fit.get_metadata_item('prop_zero_deaths'))
+        tt = tt.append(row,ignore_index=True)
+        func = np.append(func,float(fit.get_metadata_item('fn')))
+        beta = np.exp(diag['logbeta'])
+        mbeta = np.append(mbeta,stats.median(beta))
+        mu = np.exp(diag['logmu'])
+        mmu = np.append(mmu,mu.quantile(q=0.5))
+    #   gamma = diag['gamma'
+    #   mgamma = np.append(mgamma,gamma.quantile(q=0.5))
+
+    tt['fn'] = func
+#   tt['mgamma'] = mgamma
+    tt['mbeta'] = mbeta
+    tt['mmu'] = mmu
+
+    mtime = os.path.getmtime(cv.NYT_counties)
+    dtime = datetime.fromtimestamp(mtime)
+    ft_name = cv.fit_path+'fit_table_'+str(dtime.date())
+
+    csv = ft_name+'.csv'
+    tt.to_csv(csv,index=False)
+    print('Fit table data written to file',csv)
+
+
+    tt = tt.sort_values(by='mbeta',ascending=True)#,inplace=True)
+
+    for r in range(0,tt.shape[0]):
+        for c in range(3,len(tt.columns)):
+           if (tt.iloc[r,c] != None):
+               tt.iloc[r,c] = round(float(tt.iloc[r,c]),sigfigs)
+        c = 2
+        tt.iloc[r,c] = round(float(tt.iloc[r,c]),sigfigs)
+
+    row = pd.Series(None,index=tt.columns)
+    row['county'] = 'Median'
+    for n in tt.columns:
+        if (n != 'county'):
+            mn = tt[n].quantile()
+            row[n] = mn
+    tt = tt.append(row,ignore_index=True)
+
+
+    tex = ft_name+'.tex'
+    ff = open(tex, 'w')
+    caption_text = "Model results. Estimating $\\beta$ and $\mu$ trends as random effects with $\gamma = 0$.\nData updated " + str(dtime.date()) + " from https://github.com/nytimes/covid-19-data.git.\n"
+
+    ff.write(caption_text)
+#   ff.write(str(dtime.date())+'\n')
+    ff.write(tabulate(tt, header, tablefmt="latex_raw",showindex=False))
+#   tt.to_latex(buf=tex,index=False,index_names=False,longtable=False,
+#               header=header,escape=False,#float_format='{:0.4f}'.format
+#               na_rep='',column_format='lrrrrrrrrrrr')
+    print('Fit table written to file',tex)
